Moved ResponseOptionsImpl to a  top-level class
